Revert "Refatoração p/ atividade de Engenharia de Software 2"
This reverts commit cc258bdec8248d5ba9a951eea2efc6ccbd42216c.
